Fix tail bug
The bug was that when the channel was explicitly specified, it always
evaluated to stderr, i.e. supervisorctl tail foo stdout and
supervisorctl tail foo stderr were both showing stderr.
That is because we should not use is for such comparisons:
these two are different objects.

If curious, read more here: https://stackoverflow.com/questions/1504717/

This just changes the tail command to use == for comparing the two objects.