Revert "Fix DeprecationWarning on Python 3.8"
Fixes #1276.  Differences in behavior between the deprecated split
functions and urlparse() cause this to fail under some conditions.
For now, revert the changes.  A new patch will be needed to fix
the Python 3.8 warnings.  It is better to have it work with
warnings on 3.8 than to not work on many versions.

This reverts commit b13a0055aba6f665dee0d782a388082463e7a949.