Correct environment= docs, add test for actual behavior
It was observed in #550 that when environment= is set in the
[supervisord] section of the config file, that the environment
of the supervisord process itself was never actually changed.
This was contrary to the documentation, which said it would be.
It has been this way for a very long time (always?).

Now the documentation has been corrected to indicate that
environment= in the [supervisord] section will be inherited by
all subprocesses, but will not actually change the environment
of supervisord itself.  This may actually be safer, as changing
some environment variables such as PYTHONPATH could unexpectedly
affect the execution of supervisord (unsure).

Closes #550